feat(column mapper): create dynamic column options
In order to use MySQL and SQLITE3, a column mapper was created to have dynamic column options
available to run the application and its tests

BREAKING CHANGE: Updated column options

feat #167
